--- src/app/block/block.component.ts
@@ -1,12 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-block',
-  standalone: true,
-  imports: [],
-  templateUrl: './block.component.html',
-  styleUrl: './block.component.scss'
-})
-export class BlockComponent {
-
-}
