+++ src/app/components/block/block.component.html
@@ -1,0 +1,14 @@
+<div class="h-full p-4">
+    <div class="h-1/2 m-1 border rounded-3xl shadow-lg bg-transparent">
+        <h1 class="text-2xl font-extrabold text-transparent text-center bg-gradient-to-r from-pink-800 to-red-200 bg-clip-text stroke-white stroke-2 m-2">
+            POS1
+        </h1>
+    </div>
+    <div class="h-1/2 m-1 border rounded-3xl shadow-lg bg-transparent">
+        <h1 class="text-2xl font-extrabold text-transparent text-center bg-gradient-to-r from-pink-800 to-red-200 bg-clip-text stroke-white stroke-2 m-2">
+            POS2
+        </h1>
+    </div>
+</div>
+
+
